apps/api/requirements.txt
# Web Framework
python-dotenv==1.1.0
fastapi==0.110.0
uvicorn==0.22.0
python-multipart>=0.0.9

# Database
sqlmodel==0.0.24
SQLAlchemy==2.0.40
alembic==1.13.1
psycopg2-binary==2.9.10 # Use pre-compiled binary version

# Testing
pytest>=8.0.0
pytest-mock>=3.12.0
httpx==0.24.1

# Utilities
colorlog==6.9.0
gevent==24.11.1
requests==2.32.4